@@ -1,6 +1,6 @@
 {
   "name": "@signalk/server-admin-ui-dependencies",
-  "version": "1.0.1",
+  "version": "2.22.0",
   "description": "UI dependencies of the admin UI exposed to plugins that expose federated components",
   "main": "index.js",
   "author": "teppo.kurki@iki.fi",
@@ -10,15 +10,12 @@
   },
   "peerDependencies": {
     "@fortawesome/fontawesome-free": "^5.15.1",
-    "bootstrap": "^4.5.3",
+    "bootstrap": "^5.3.3",
     "font-awesome": "^4.7.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-redux": "^5.1.2",
-    "react-select": "^3.1.0",
-    "reactstrap": "^5.0.0",
-    "redux": "^3.7.2",
-    "redux-thunk": "2.3.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-bootstrap": "^2.10.10",
+    "react-select": "^5.10.2",
     "simple-line-icons": "^2.5.5"
   }
 }
